# RentalManager.WebApp/src/references/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RentalManager.WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Client/{id}": {
      "delete": {
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/Employee/{id}": {
      "delete": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployee"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployee"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/Payment/{id}": {
      "delete": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePayment"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePayment"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePayment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/RentalAgreement/{id}": {
      "delete": {
        "tags": [
          "RentalAgreement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "RentalAgreement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RentalAgreement"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalAgreement"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalAgreement"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalAgreement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalAgreementDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "idCard": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "RentalAgreementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "employee": {
            "$ref": "#/components/schemas/EmployeeDto"
          },
          "isActive": {
            "type": "boolean"
          },
          "client": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "rentalEquipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalEquipmentDto"
            },
            "nullable": true
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            },
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "deposit": {
            "type": "integer",
            "format": "int32"
          },
          "transportFrom": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "transportTo": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "RentalEquipmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateClient": {
        "required": [
          "city",
          "name",
          "phoneNumber",
          "street",
          "surname"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "default": "John"
          },
          "surname": {
            "minLength": 1,
            "type": "string",
            "default": "Kowalski"
          },
          "phoneNumber": {
            "minLength": 1,
            "pattern": "^((\\+[0-9]{1,4})?[-. ]?([0-9]{3})[-. ]?([0-9]{3})[-. ]?([0-9]{3}))$|^(\\(?[0-9]{2}\\)?[-. ]?[0-9]{2}[-. ]?[0-9]{3}[-. ]?[0-9]{2})$",
            "type": "string",
            "default": "+48 123 123 123"
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-Z]*\\@[a-zA-Z]{3,}\\.[a-zA-Z]{2,5}$",
            "type": "string",
            "default": "JohnKowalski@email.com",
            "nullable": true
          },
          "idCard": {
            "pattern": "^([a-zA-Z]){3}[ ]?[0-9]{6}$",
            "type": "string",
            "default": "ABC 123456",
            "nullable": true
          },
          "city": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ]*$",
            "type": "string",
            "default": "Warsaw"
          },
          "street": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ 0-9\\/.]*$",
            "type": "string",
            "default": "Wiejska 4/6/8"
          }
        },
        "additionalProperties": false
      },
      "UpdateEmployee": {
        "required": [
          "name",
          "surname"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ]*$",
            "type": "string",
            "default": "John"
          },
          "surname": {
            "maxLength": 100,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ]*$",
            "type": "string",
            "default": "Kowalski"
          }
        },
        "additionalProperties": false
      },
      "UpdatePayment": {
        "type": "object",
        "properties": {
          "method": {
            "maxLength": 10,
            "minLength": 0,
            "pattern": "^[a-zA-ZąćęłńóśźżĄĘŁŃÓŚŹŻ]*$",
            "type": "string",
            "default": "Cash",
            "nullable": true
          },
          "amount": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 100
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateRentalAgreement": {
        "type": "object",
        "properties": {
          "employeeId": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "clientId": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "rentalEquipmentIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "default": [
              1
            ],
            "nullable": true
          },
          "comment": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "default": "Fun comment",
            "nullable": true
          },
          "deposit": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 100
          },
          "transportFrom": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 100,
            "nullable": true
          },
          "transportTo": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}